# Geometry2D minor optimization